+++ src/index.ts
@@ -57,7 +57,7 @@ import type {Handler, HandlerResult} from './ui/handler_manager.ts';
 import type {Complete, Mat4f32, Mat4f64, RequireAtLeastOne, Subscription} from './util/util.ts';
 import type {CalculateTileZoomFunction, CoveringTilesOptions} from './geo/projection/covering_tiles.ts';
 import type {TransformConstrainFunction} from './geo/transform_interface.ts';
-import type {CustomStyleImageInterface, StyleImage, StyleImageContext, StyleImageData, StyleImageInterface, StyleImageMetadata, StyleImageRenderTarget, TextFit} from './style/style_image.ts';
+import type {StyleImage, StyleImageData, StyleImageInterface, StyleImageMetadata, StyleImageWebGLData, StyleImageWebGLTarget, TextFit} from './style/style_image.ts';
 import type {StyleLayer, PaintPropertyEntry} from './style/style_layer.ts';
 import type {Tile} from './tile/tile.ts';
 import type {GeoJSONFeatureDiff, GeoJSONFeatureId, GeoJSONSourceDiff} from './source/geojson_source_diff.ts';
@@ -308,9 +308,8 @@ export {
     type GeolocateControlOptions,
     type LogoControlOptions,
     type StyleImageInterface,
-    type CustomStyleImageInterface,
-    type StyleImageContext,
-    type StyleImageRenderTarget,
+    type StyleImageWebGLData,
+    type StyleImageWebGLTarget,
     type AddLayerObject,
     type StyleSetterOptions,
     type CameraForBoundsOptions,